update scheduler view, add functions to update job interval
