+++ webtool/__init__.py
@@ -106,6 +106,7 @@ import webtool.views.views_user
 
 import webtool.views.views_dataset
 import webtool.views.views_misc
+import webtool.views.views_scheduler
 
 import webtool.views.api_explorer
 import webtool.views.api_standalone
